major change, restructure classes, agregation and custom sql

+++ src/main/kotlin/cz/upce/nnpia/personstatistics/rest/UserController.kt
@@ -4,6 +4,7 @@ import cz.upce.nnpia.personstatistics.auth.IAuthenticationFacade
 import cz.upce.nnpia.personstatistics.dto.UserDto
 import cz.upce.nnpia.personstatistics.security.UserRole
 import cz.upce.nnpia.personstatistics.service.implementations.UserService
+import cz.upce.nnpia.personstatistics.service.interfaces.IUserService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.*
@@ -13,7 +14,7 @@ import org.springframework.web.bind.annotation.*
 @CrossOrigin()
 class UserController
 @Autowired constructor(
-	private val userService: UserService,
+	private val userService: IUserService,
 	private val authenticationFacade: IAuthenticationFacade
 ) {
 
@@ -21,7 +22,7 @@ class UserController
 	fun handleException(): String = "error"
 
 	fun getUserId(): Long {
-		return userService.getByUsername(authenticationFacade.getUserName()).id
+		return userService.getByUsername(authenticationFacade.getUserName())?.id
 			?: throw IllegalStateException("User id is not exist")
 	}
 
@@ -35,10 +36,12 @@ class UserController
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	fun remove(@PathVariable id: Long) {
 		val person = userService.getById(id)
-		if (authenticationFacade.getUserName() != person.username) {
-			userService.removeById(id)
-		} else {
-			throw IllegalStateException("Impossible delete admin account")
+		if (person != null) {
+			if (authenticationFacade.getUserName() != person.username) {
+				userService.removeById(id)
+			} else {
+				throw IllegalStateException("Impossible delete admin account")
+			}
 		}
 	}
